requirements.txt
Django==4.1
gunicorn==20.1.0
Pillow==9.2.0
whitenoise==6.4.0
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
celery==5.4.0
cffi==1.17.1
channels==4.2.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
Django==5.1.7
django-celery-results==2.5.1
django-filter==25.1
django-mongoengine==0.5.6
django-redis==5.4.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.5.1
mongoengine==0.29.1
pillow==11.1.0
prompt_toolkit==3.0.50
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pymongo==4.11.3
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
redis==5.2.1
service-identity==24.2.0
setuptools==78.1.0
six==1.17.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.0
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2